Computer / WMI & CSV
This Python script utilizes the Windows Management Interface (WMI) and the Comma Separated Value (CSV) modules, in order to query and save specific information. It then sends the file to a specific path which can be referenced later.

The reason why I like this is tool, is because it improves productivity and decreases the risk for errors. I It’s a great tool for documenting computer information, collect information, and to continue to develop even further.